migrated hasOld tests
