+++ test/unit.html
@@ -33,7 +33,6 @@
   <script src="expandKeyValue_test_old.js"></script>
   <script src="flatten_test_old.js"></script>
   <script src="get_test_old.js"></script>
-  <script src="hasOwn_test_old.js"></script>
   <script src="isArray_test_old.js"></script>
   <script src="processStates_test_old.js"></script>
 </body>
